Database and protocol implementation

+++ sourcestats_database/migrations/2019-01-24-184256_create_players/up.sql
@@ -1,4 +1,4 @@
 CREATE TABLE players (
-    player_id SERIAL PRIMARY KEY,
-    steam_id NUMERIC UNIQUE NOT NULL
+    player_id BIGSERIAL PRIMARY KEY,
+    steam_id BIGINT UNIQUE NOT NULL
 );
