+++ sourcestats_database/migrations/2019-01-24-212527_add_encryption_keys/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE server_keys (
-    key_id BIGSERIAL PRIMARY KEY,
+    key_id BIGINT PRIMARY KEY,
     key_data BYTEA NOT NULL,
-    server_id INTEGER NOT NULL REFERENCES servers(server_id)
+    server_id BIGINT NOT NULL REFERENCES servers(server_id)
 );
